Change ValidateRange from int to long, add check for json file not existing for new RPs

+++ tools/StaticAnalysis/BreakingChangeAnalyzer/ParameterMetadata.cs
@@ -50,13 +50,13 @@ namespace StaticAnalysis.BreakingChangeAnalyzer
         /// If the parameter implements ValidateRange, specifies the minimum value
         /// in the range; otherwise, the value is null
         /// </summary>
-        public int? ValidateRangeMin { get; set; }
+        public long? ValidateRangeMin { get; set; }
 
         /// <summary>
         /// If the parameter implements ValidateRange, specifies the maximum value
         /// in the range; otherwise, the value is null
         /// </summary>
-        public int? ValidateRangeMax { get; set; }
+        public long? ValidateRangeMax { get; set; }
 
         /// <summary>
         /// Specifies if the parameter has the ValidateNotNullOrEmpty attribute
